+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## 2.34.0 - 2022-07-05
+### Added
+- Support for backup policies returned as part of the database system list operation in the MySQL Database service
+ 
+### Breaking Changes
+- Default retry enable for operations of the Bastion service
+
 ## 2.33.0 - 2022-06-27
 ### Added
 - Support for the Network Monitoring service